Turn off 'make-backup-files behavior
This is being done because UNIX file watch is observing the inode & not
the path to the file. So, inotify and other approaches to generating an
event "on file change" are tricked because the "watched" file is
passively copied to a _backup_ and a new file modified. (ref: [0,1])

This is likely why webpack-dev-server "watch" mode has stopped working
for me when I moved to emacs from Sublime Text.

[0] https://github.com/vickychijwani/inotify-file-watcher
[1] https://github.com/vickychijwani/inotify-file-watcher/blob/7eb635466b120dd2850a3f81aacc25b97529ddb2/README.md#caveats